Abstract SkinStorage for adding adapters easier in the future
